refactor(api): rename ui subpackage name
