+++ packages/api/jsr.json
@@ -6,8 +6,8 @@
 	"exports": {
 		".": "./src/index.ts",
 		"./ui": "./src/ui/index.ts",
-		"./ui/iframe": "./src/ui/iframe/index.ts",
-		"./ui/worker": "./src/ui/worker/index.ts",
+		"./ui/custom": "./src/ui/custom/index.ts",
+		"./ui/template": "./src/ui/template/index.ts",
 		"./headless": "./src/headless/index.ts",
 		"./models": "./src/models/index.ts",
 		"./commands": "./src/commands/index.ts",
