+++ packages/api/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@kksh/api",
-	"version": "0.1.0",
+	"version": "0.1.1",
 	"type": "module",
 	"repository": {
 		"type": "git",
@@ -12,8 +12,8 @@
 	"exports": {
 		".": "./src/index.ts",
 		"./ui": "./src/ui/index.ts",
-		"./ui/iframe": "./src/ui/iframe/index.ts",
-		"./ui/worker": "./src/ui/worker/index.ts",
+		"./ui/custom": "./src/ui/custom/index.ts",
+		"./ui/template": "./src/ui/template/index.ts",
 		"./headless": "./src/headless/index.ts",
 		"./models": "./src/models/index.ts",
 		"./commands": "./src/commands/index.ts",
