+++ packages/api/src/ui/template/components/common.ts
