+++ packages/api/src/ui/template/components/icon.ts
