+++ packages/api/src/ui/template/schema/index.ts
