+++ packages/api/src/ui/template/schema/keyboard.ts
