+++ packages/templates/template-ext-react/src/App.tsx
@@ -1,4 +1,4 @@
-import { ui } from "@kksh/api/ui/iframe"
+import { ui } from "@kksh/api/ui/custom"
 import {
 	ActionPanel,
 	Button,
