+++ packages/templates/template-ext-sveltekit/src/lib/components/ThemeCustomizer.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { ThemeCustomizerButton, type ThemeConfig, updateTheme } from '@kksh/svelte5';
-	import { ui } from '@kksh/api/ui/iframe';
+	import { ui } from '@kksh/api/ui/custom';
 	import { onMount } from 'svelte';
 
 	let config: ThemeConfig = {
