+++ packages/templates/template-ext-sveltekit/src/routes/+layout.svelte
@@ -3,7 +3,7 @@
 	import { ModeWatcher } from 'mode-watcher';
 	import { ThemeWrapper, updateTheme } from '@kksh/svelte5';
 	import { onMount } from 'svelte';
-	import { ui } from '@kksh/api/ui/iframe';
+	import { ui } from '@kksh/api/ui/custom';
 
 	onMount(() => {
 		ui.registerDragRegion();
