+++ packages/templates/template-ext-worker/src/index.ts
@@ -12,7 +12,7 @@ import {
 	TemplateUiCommand,
 	toast,
 	ui
-} from "@kksh/api/ui/worker"
+} from "@kksh/api/ui/template"
 
 class ExtensionTemplate extends TemplateUiCommand {
 	async onFormSubmit(value: Record<string, any>): Promise<void> {
